extract sendLocation into its own file createLocation
